move some components to atoms
