--- src/components/organisms/Backdrop.js
@@ -1,9 +1,0 @@
-import classes from './Backdrop.module.css';
-
-const Backdrop = (props) => {
-  return (
-    <div className={classes['backdrop']} onClick={props.onSideMenuClose}></div>
-  );
-};
-
-export default Backdrop;
